backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
openai
httpx<0.28.0
groq
numpy
